fixed mask issues in tests (#55), move the validation of baseline into the explainer file
