+++ src/meteors/attr/integrated_gradients.py
@@ -10,34 +10,7 @@ from meteors.utils.models import ExplainableModel
 from meteors import Image
 from meteors.attr import ImageAttributes, Explainer
 
-## VALIDATORS
-
-
-# @lru_cache(maxsize=32)
-def validate_and_transform_baseline(baseline: int | float | torch.Tensor | None, image: Image) -> torch.Tensor:
-    """Function validates the baseline and transforms it to the same device as the image tensor.
-
-    Args:
-        baseline (int | float | torch.Tensor): _description_
-        image (Image): an Image object for which the baseline is being validated
-
-    Raises:
-        ValueError: _description_
-
-    Returns:
-        torch.Tensor: _description_
-    """
-
-    if baseline is None:
-        baseline = 0
-    if isinstance(baseline, (int, float)):
-        baseline = torch.zeros_like(image.image) + baseline
-    elif isinstance(baseline, torch.Tensor):
-        if baseline.shape != image.image.shape:
-            raise ValueError(f"Baseline shape {baseline.shape} does not match image shape {image.image.shape}")
-
-    baseline = baseline.to(image.image.device)  # cast the baseline to the same device as the image tensor
-    return baseline
+from meteors.attr.explainer import validate_and_transform_baseline
 
 
 class IntegratedGradients(Explainer):
